% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sv.R
\name{plot_circos}
\alias{plot_circos}
\title{Plot SVs}
\usage{
plot_circos(manta, circos_path, outdir, outfile = "circos.png",
  genome = "hg38", filter_pass = TRUE)
}
\arguments{
\item{manta}{Path to \code{sv.vcf.gz} file.}

\item{circos_path}{Path to the directory containing the circos binary (see examples).}

\item{outdir}{Directory to write the output circos file along with the configs used.}

\item{outfile}{Name of PNG file.}

\item{genome}{Genome assembly used to generate the SV calls.}

\item{filter_pass}{Logical. Should we keep only the PASSed variants?}
}
\value{
Path to circos PNG file output by the Perl circos program.
}
\description{
Plots the SVs found in the \code{sv.vcf.gz} VCF file output by DRAGEN.
}
\examples{
\dontrun{
circos_png <- plot_circos(
   manta = "path/to/sample.sv.vcf",
   circos_path = "path/to/bin_with_circos",
   outdir = file.path("circos"))
}
}
