% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fragment_length.R
\name{plot_fragment_length_hist}
\alias{plot_fragment_length_hist}
\title{Plot insert length distribution}
\usage{
plot_fragment_length_hist(x, min_count = 10)
}
\arguments{
\item{x}{Path to \code{fragment_length_hist.csv} file.}

\item{min_count}{Minimum read count to be plotted (Default: 10).}
}
\value{
A ggplot2 plot containing insert lengths on X axis and read counts
on Y axis for each sample.
}
\description{
Plots the insert length distributions as written in the \code{fragment_length_hist.csv} file.
}
\examples{
x <- system.file("extdata/Colo829.fragment_length_hist.csv.gz", package = "dracarys")
p <- plot_fragment_length_hist(x, min_count = 10)
}
