% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sv.R
\name{plot_sv_contig_counts}
\alias{plot_sv_contig_counts}
\title{Plot SV Counts per Contig}
\usage{
plot_sv_contig_counts(manta)
}
\arguments{
\item{manta}{Path to \code{sv.vcf.gz} file or object of class 'sv'.}
}
\value{
A ggplot2 object plotting number of SV breakpoints per contig for
mate1 and mate2 of each SV event.
}
\description{
Plots number of SV breakpoints per contig for mate1 and mate2 of each
SV event.
}
\examples{
manta <- system.file("extdata/COLO829.sv.vcf.gz", package = "dracarys")
plot_sv_contig_counts(manta)

manta_table <- dracarys::sv_table(manta)
plot_sv_contig_counts(manta_table)

}
