% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qc_coverage.R
\name{plot_wgs_contig_coverage}
\alias{plot_wgs_contig_coverage}
\title{Plot WGS Contig Mean Coverage Files}
\usage{
plot_wgs_contig_coverage(xs, labels, top_alt_n = 15)
}
\arguments{
\item{xs}{Character vector containing paths to \code{wgs_contig_mean_cov_<phenotype>.csv} files.}

\item{labels}{Character vector containing labels for the files specified in \code{xs}.}

\item{top_alt_n}{Number of top covered alt contigs to plot per phenotype.}
}
\value{
A ggplot2 object with chromosomes on X axis, and coverage on Y axis.
}
\description{
Plots the \code{wgs_contig_mean_cov_<phenotype>.csv} files.
}
\examples{
normal <- system.file("extdata/COLO829.wgs_contig_mean_cov_normal.csv.gz", package = "dracarys")
tumor <- system.file("extdata/COLO829.wgs_contig_mean_cov_tumor.csv.gz", package = "dracarys")

plot_wgs_contig_coverage(xs = c(tumor, normal), labels = c("tumor", "normal"))
}
