% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qc_coverage.R
\name{plot_wgs_fine_cumsum}
\alias{plot_wgs_fine_cumsum}
\title{Plot WGS Fine Hist File Cumsum}
\usage{
plot_wgs_fine_cumsum(tumor, normal, colours = c("#56B4E9", "#D55E00"),
  y_lim = 0.003)
}
\arguments{
\item{tumor}{Path to \code{wgs_fine_hist_tumor.csv} file.}

\item{normal}{Path to \code{wgs_fine_hist_normal.csv} file.}

\item{colours}{Colours for normal and tumor sample, in that order.}

\item{y_lim}{Y axis lower limit.}
}
\value{
A ggplot2 object with depth of coverage on X axis, and fraction
of loci with that depth on Y axis.
}
\description{
Plots the \code{wgs_fine_hist_<phenotype>.csv} coverage cumsum.
}
\examples{
normal <- system.file("extdata/COLO829.wgs_fine_hist_normal.csv.gz", package = "dracarys")
tumor <- system.file("extdata/COLO829.wgs_fine_hist_tumor.csv.gz", package = "dracarys")

plot_wgs_fine_cumsum(tumor = tumor, normal = normal)
}
