% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fragment_length.R
\name{read_fragment_length_hist}
\alias{read_fragment_length_hist}
\title{Read fragment length hist file}
\usage{
read_fragment_length_hist(x)
}
\arguments{
\item{x}{Path to \code{fragment_length_hist.csv} file.}
}
\value{
A tibble with the following columns:
\itemize{
\item fragmentLength
}
}
\description{
Reads the \code{fragment_length_hist.csv} file, which contains the
insert length distribution for each sample.
}
\examples{
x <- system.file("extdata/Colo829.fragment_length_hist.csv.gz", package = "dracarys")
read_fragment_length_hist(x)
}
