% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qc_mapping_metrics.R
\name{read_mapping_metrics}
\alias{read_mapping_metrics}
\title{Read Mapping Metrics File}
\usage{
read_mapping_metrics(x)
}
\arguments{
\item{x}{Path to \code{mapping_metrics.csv} file.}
}
\value{
A list containing two elements:
\itemize{
\item data: tibble with the following columns:
\itemize{
\item category: summary or read group
\item Phenotype: e.g. tumor, normal
\item RG: read group
\item var: metric variable
\item count: count of reads
\item pct: percentage of reads
}
\item abbreviations: named character vector with column abbreviations
}
}
\description{
Reads the \code{mapping_metrics.csv} file, which contains
mapping and aligning metrics, like the metrics computed by the Samtools
Flagstat command. These metrics are available on an aggregate level
(over all input data), and on a per read group level.
Unless explicitly stated, the metrics units are in reads (i.e., not in
terms of pairs or alignments).
}
\examples{
x <- system.file("extdata/COLO829.mapping_metrics.csv.gz", package = "dracarys")
read_mapping_metrics(x)
}
