% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qc_ploidy_estimation_metrics.R
\name{read_ploidy_estimation_metrics}
\alias{read_ploidy_estimation_metrics}
\title{Read Ploidy Estimation Metrics File}
\usage{
read_ploidy_estimation_metrics(x)
}
\arguments{
\item{x}{Path to \code{ploidy_estimation_metrics.csv} file.}
}
\value{
tibble with the following columns:
- category: summary or read group
- Phenotype: e.g. tumor, normal
- RG: read group
- var: metric variable
- var_abbrev: metric variable abbreviation
- count: count of reads
- pct: percentage of reads
}
\description{
Reads the \code{ploidy_esimation_metrics.csv} file
}
\examples{
x <- system.file("extdata/COLO829.ploidy_estimation_metrics.csv.gz", package = "dracarys")
read_ploidy_estimation_metrics(x)
}
