% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dragen_version.R
\name{read_replay}
\alias{read_replay}
\title{Read Replay File}
\usage{
read_replay(x, label)
}
\arguments{
\item{x}{Path to \code{replay.json} file.}

\item{label}{Label for the file.}
}
\value{
A list with the following elements:
\itemize{
\item \code{command_line}: character. DRAGEN command line used.
\item \code{dragen_config}: tibble. Parameters used for the DRAGEN run.
\item \code{inputs}: tibble. Germline input FASTQ files (if run in Tumor/Normal mode).
\item \code{system}: list. DRAGEN version, node name, and kernel release.
}
}
\description{
Reads the \code{replay.json} file, which contains the DRAGEN command line,
parameters, version and inputs for the specific run.
}
\examples{
x <- system.file("extdata/COLO829-replay.json.gz", package = "dracarys")
read_replay(x, label = "COLO829")

}
