% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/roh.R
\name{read_roh_metrics}
\alias{read_roh_metrics}
\title{Read ROH Metrics File}
\usage{
read_roh_metrics(x)
}
\arguments{
\item{x}{Path to \code{roh_metrics.csv} file.}
}
\value{
A tibble with two columns:
- \code{pct_snv_in_3mb_roh}: percent of SNVs in large (>= 3Mb) ROH regions.
- \code{n_3mb_roh}: number of large (>= 3Mb) ROH regions.
}
\description{
Read ROH Metrics File
}
\examples{
x <- system.file("extdata/COLO829_N.roh_metrics.csv", package = "dracarys")
(roh <- read_roh_metrics(x))

}
