% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qc_runtime_metrics.R
\name{read_time_metrics}
\alias{read_time_metrics}
\title{Read Time Metrics File}
\usage{
read_time_metrics(x)
}
\arguments{
\item{x}{Path to \code{time_metrics.csv} file.}
}
\value{
tibble with the following columns:
\itemize{
\item Step: DRAGEN step
\item Time: time in HH:MM:SS
}
}
\description{
Reads the \code{time_metrics.csv} file, which contains
a breakdown of the run duration for each DRAGEN process.
}
\examples{
x <- system.file("extdata/COLO829.time_metrics.csv.gz", package = "dracarys")
(tm <- read_time_metrics(x))

}
