% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qc_coverage.R
\name{read_wgs_contig_coverage}
\alias{read_wgs_contig_coverage}
\title{Read WGS Contig Mean Coverage File}
\usage{
read_wgs_contig_coverage(x, keep_alt = FALSE)
}
\arguments{
\item{x}{Path to \verb{wgs_contig_mean_cov_<phenotype>.csv} file.}

\item{keep_alt}{Keep the ALT + Mito chromosomes?}
}
\value{
tibble with following columns:
\itemize{
\item label
\item chrom
\item n_bases
\item coverage
}
}
\description{
Reads the \verb{wgs_contig_mean_cov_<phenotype>.csv} file, which contains
the estimated coverage for all contigs, and an autosomal estimated coverage.
Columns are:
\itemize{
\item contig name.
\item number of bases aligned to contig
(excludes bases from duplicate marked reads, reads with MAPQ=0, and clipped bases).
\item coverage estimate = col2 / contig length
}
}
\examples{
x <- system.file("extdata/COLO829.wgs_contig_mean_cov_normal.csv.gz", package = "dracarys")
y <- system.file("extdata/COLO829.wgs_contig_mean_cov_tumor.csv.gz", package = "dracarys")

(cc_x <- read_wgs_contig_coverage(x))
(cc_y <- read_wgs_contig_coverage(y))

}
