% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qc_coverage.R
\name{read_wgs_coverage_metrics}
\alias{read_wgs_coverage_metrics}
\title{Read WGS Coverage Metrics File}
\usage{
read_wgs_coverage_metrics(x)
}
\arguments{
\item{x}{Path to \verb{wgs_coverage_metrics_<phenotype>.csv} file.}
}
\value{
tibble with following columns:
\itemize{
\item label: file label.
\item var: variable name.
\item var_abbrev: variable abbreviation.
\item pct: percentage value.
\item count: count value.
}
}
\description{
Reads the \verb{wgs_coverage_metrics_<phenotype>.csv} file, which contains
read depth of coverage metrics.
}
\examples{
x <- system.file("extdata/COLO829.wgs_coverage_metrics_normal.csv.gz", package = "dracarys")
y <- system.file("extdata/COLO829.wgs_coverage_metrics_tumor.csv.gz", package = "dracarys")

(cm_x <- read_wgs_coverage_metrics(x))
(cm_y <- read_wgs_coverage_metrics(y))

}
