% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qc_coverage.R
\name{read_wgs_fine_hist}
\alias{read_wgs_fine_hist}
\title{Read WGS Fine Hist File}
\usage{
read_wgs_fine_hist(x, label)
}
\arguments{
\item{x}{Path to \code{wgs_fine_hist_<phenotype>.csv} file.}

\item{label}{Label for the file e.g. 'tumor' or 'normal'.}
}
\value{
tibble with two columns:
\itemize{
\item Depth
\item Overall
}
}
\description{
Reads the \code{wgs_fine_hist_<phenotype>.csv} file, which contains
two columns: Depth and Overall.
The value in the Depth column ranges from 0 to 1000+ and the Overall
column indicates the number of loci covered at the corresponding depth.
}
\examples{
x <- system.file("extdata/COLO829.wgs_fine_hist_normal.csv.gz", package = "dracarys")
y <- system.file("extdata/COLO829.wgs_fine_hist_tumor.csv.gz", package = "dracarys")

read_wgs_fine_hist(x, label = "normal")
read_wgs_fine_hist(x, label = "tumor")
}
