% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/roh.R
\name{rohbed2circos}
\alias{rohbed2circos}
\title{Read ROH BED File}
\usage{
rohbed2circos(x)
}
\arguments{
\item{x}{Path to \code{roh.bed} file.}
}
\value{
A tibble with 4 columns:
- \code{chrom}, \code{start}, \code{end}: coordinates for one region of homozygosity (0-based, half-open interval).
The \code{chrom} is returned with a \code{hs} prefix instead of \code{chr}.
- \code{score}: function of the number of homozygous and heterozygous variants,
where each homozygous variant increases the score by a pre-defined value,
and each heterozygous variant reduces the score by (1 - pre-defined value),
where the pre-defined value is in range (0, 1).
}
\description{
Read ROH BED File
}
\examples{
x <- system.file("extdata/COLO829_N.roh.bed", package = "dracarys")
(roh <- rohbed2circos(x))

}
