% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setup_vars.R
\name{setup_vars}
\alias{setup_vars}
\title{Set Up dracarys Variables}
\usage{
setup_vars(x)
}
\arguments{
\item{x}{Path to \code{replay.json} file.}
}
\value{
A list with several elements pointing to DRAGEN output file names.
}
\description{
Reads the \code{replay.json} file, which contains the DRAGEN command line,
parameters, version and inputs for the specific run. It then pre-creates
expected file names for DRAGEN output.
}
\examples{
x <- system.file("extdata/COLO829-replay.json.gz", package = "dracarys")
setup_vars(x)

}
