% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/snv.R
\name{snv_table}
\alias{snv_table}
\title{Generate Table with Single Nucleotide/INDEL Variants}
\usage{
snv_table(x, prefix, outdir, env_path)
}
\arguments{
\item{x}{Path to \code{hard-filtered.vcf.gz} file.}

\item{prefix}{Prefix for output variant and sample files.}

\item{outdir}{Output directory for results.}

\item{env_path}{Path to the directory containing the python environment
for snv_table.}
}
\value{
A list with two elements:
- A tibble with SNV/INDEL variants.
- A tibble with sample names corresponding to the VCF columns.
}
\description{
Generate Table with Single Nucleotide/INDEL Variants
}
\examples{
env_path <- Sys.getenv("DRACARYS_ENV")
x <- system.file("extdata/COLO829.hard-filtered.vcf.gz", package = "dracarys")
(snvs <- snv_table(x, prefix = "COLO829", outdir = tempdir(), env_path = env_path))

}
