% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qc_runtime_metrics.R
\name{time_metrics_process}
\alias{time_metrics_process}
\title{Process Time Metrics File}
\usage{
time_metrics_process(x, id = seq_len(length(x)))
}
\arguments{
\item{x}{Path to \code{time_metrics.csv} file. It can take multiple inputs (as a character vector).}

\item{id}{ID for each input (used to disambiguate files run on same samples). Default: index from 1 to length of \code{x}.}
}
\value{
tibble with the following columns:
\itemize{
\item Step: DRAGEN step
\item Time: time in HH:MM
}
}
\description{
Processes the \code{time_metrics.csv} file.
}
\examples{
x <- system.file("extdata/COLO829.time_metrics.csv.gz", package = "dracarys")
(tm <- time_metrics_process(c(x, x), id = c("run1", "run2")))

}
