% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qc_coverage.R
\name{plot_wgs_contig_coverage}
\alias{plot_wgs_contig_coverage}
\title{Plot WGS Contig Mean Coverage Files}
\usage{
plot_wgs_contig_coverage(tumor, normal, colours = c("#56B4E9",
  "#D55E00"))
}
\arguments{
\item{tumor}{Path to \code{wgs_contig_mean_cov_tumor.csv} file.}

\item{normal}{Path to \code{wgs_contig_mean_cov_normal.csv} file.}

\item{colours}{Colours for normal and tumor sample, in that order.}
}
\value{
A ggplot2 object with chromosomes on X axis, and coverage on Y axis.
}
\description{
Plots the \code{wgs_contig_mean_cov_<phenotype>.csv} files for tumor and normal.
}
\examples{
normal <- system.file("extdata/COLO829.wgs_contig_mean_cov_normal.csv.gz", package = "dracarys")
tumor <- system.file("extdata/COLO829.wgs_contig_mean_cov_tumor.csv.gz", package = "dracarys")

plot_wgs_contig_coverage(tumor = tumor, normal = normal)
}
