% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qc_coverage.R
\name{plot_wgs_fine_hist}
\alias{plot_wgs_fine_hist}
\title{Plot WGS Fine Hist File}
\usage{
plot_wgs_fine_hist(xs, x_lim = c(0, 300))
}
\arguments{
\item{xs}{Character vector containing paths to \code{wgs_fine_hist_<phenotype>.csv} files.}

\item{x_lim}{X axis range to plot.}
}
\value{
A ggplot2 object with depth of coverage on X axis, and number of loci with
that depth on Y axis.
}
\description{
Plots the \code{wgs_fine_hist_<phenotype>.csv} files.
}
\examples{
normal <- system.file("extdata/COLO829.wgs_fine_hist_normal.csv.gz", package = "dracarys")
tumor <- system.file("extdata/COLO829.wgs_fine_hist_tumor.csv.gz", package = "dracarys")

plot_wgs_fine_hist(xs = c(tumor, normal))
plot_wgs_fine_hist(xs = c(tumor, normal), x_lim = c(0, 500))
}
