% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qc_varcalling_metrics.R
\name{read_varcalling_metrics}
\alias{read_varcalling_metrics}
\title{Read Variant Calling Metrics File}
\usage{
read_varcalling_metrics(x)
}
\arguments{
\item{x}{Path to \code{vc_metrics.csv} file.}
}
\value{
tibble with the following columns:
}
\description{
Reads the \code{vc_metrics.csv} file, which contains variant calling metrics.
reported for each sample in multi sample VCF and gVCF files.
Based on the run case, metrics are reported either as standard VARIANT
CALLER or JOINT CALLER. Metrics are reported both for the raw
(PREFILTER) and hard filtered (POSTFILTER) VCFs.
PON (Panel of Normals) and COSMIC filtered variants are counted as
though they are PASS variants in the POSTFILTER VCFs metrics,
which may result in higher than expected variant counts in the
POSTFILTER VCF metrics.
}
\examples{
x <- system.file("extdata/COLO829.vc_metrics.csv.gz", package = "dracarys")
read_varcalling_metrics(x)
}
