% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qc_coverage.R
\name{read_wgs_contig_coverage}
\alias{read_wgs_contig_coverage}
\title{Read WGS Contig Mean Coverage File}
\usage{
read_wgs_contig_coverage(x, phenotype, keep_alt = FALSE)
}
\arguments{
\item{x}{Path to \code{wgs_contig_mean_cov_<phenotype>.csv} file.}

\item{phenotype}{Phenotype of file e.g. 'tumor', 'normal'.}

\item{keep_alt}{Keep the ALT + Mito chromosomes?}
}
\value{
tibble with following columns:
\itemize{
\item phenotype
\item chrom
\item n_bases
\item coverage
}
}
\description{
Reads the \code{wgs_contig_mean_cov_<phenotype>.csv} file, which contains
the estimated coverage for all contigs, and an autosomal estimated coverage.
Columns are:
\itemize{
\item contig name.
\item number of bases aligned to contig
(excludes bases from duplicate marked reads, reads with MAPQ=0, and clipped bases).
\item coverage estimate = col2 / contig length
}
}
\examples{
x <- system.file("extdata/COLO829.wgs_contig_mean_cov_normal.csv.gz", package = "dracarys")
y <- system.file("extdata/COLO829.wgs_contig_mean_cov_tumor.csv.gz", package = "dracarys")

read_wgs_contig_coverage(x, phenotype = "normal")
read_wgs_contig_coverage(y, phenotype = "tumor")
}
